+++ scim2_tester/checkers/patch_replace.py
@@ -6,9 +6,9 @@ from scim2_client import SCIMClientException
 from scim2_models import Mutability
 from scim2_models import PatchOp
 from scim2_models import PatchOperation
+from scim2_models import Path
 from scim2_models import Resource
 from scim2_models import SCIMException
-from scim2_models.path import Path
 
 from ..filling import generate_random_value
 from ..utils import CheckContext
@@ -45,8 +45,8 @@ def check_replace_attribute(
        specified by the 'path'."
     """
     if (
-        context.client.service_provider_config
-        and not context.client.service_provider_config.patch.supported
+        context.client.provider.config
+        and not context.client.provider.config.patch.supported
     ):
         return [
             check_result(
@@ -80,7 +80,8 @@ def check_replace_attribute(
     for path in all_paths:
         urn = str(path)
         patch_value = generate_random_value(context, path=path)
-        mutability = path.get_annotation(Mutability)
+        binding = path.resolve()
+        mutability = binding.get_annotation(Mutability) if binding else None
 
         patch_op = PatchOp[type(base_resource)](
             operations=[